# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Score My Recipe",
    "description": "A tool to compute Green-Score of recipes.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/parse_text": {
      "post": {
        "summary": "Parse Text",
        "description": "Parse a text and return a list of ingredients with quantities and eventual modifiers",
        "operationId": "parse_text_v1_parse_text_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeParseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeParseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/origins": {
      "get": {
        "summary": "Get Origins",
        "description": "Get the list of origins available in the database\n\nNote: as the list is not too big, we let clients handle suggestions to users",
        "operationId": "get_origins_v1_origins_get",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Language for the request (2 or 5 letter code)",
              "title": "Lang"
            },
            "description": "Language for the request (2 or 5 letter code)"
          },
          {
            "name": "include_synonyms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, include the synonyms of each item in the response.",
              "default": false,
              "title": "Include Synonyms"
            },
            "description": "If true, include the synonyms of each item in the response."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OriginsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/labels": {
      "get": {
        "summary": "Get Labels",
        "description": "Get the list of labels relevant for green-score computation",
        "operationId": "get_labels_v1_labels_get",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Language for the request (2 or 5 letter code)",
              "title": "Lang"
            },
            "description": "Language for the request (2 or 5 letter code)"
          },
          {
            "name": "include_synonyms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, include the synonyms of each item in the response.",
              "default": false,
              "title": "Include Synonyms"
            },
            "description": "If true, include the synonyms of each item in the response."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/countries": {
      "get": {
        "summary": "Get Countries",
        "description": "Get the list of countries relevant for green-score computation",
        "operationId": "get_countries_v1_countries_get",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Language for the request (2 or 5 letter code)",
              "title": "Lang"
            },
            "description": "Language for the request (2 or 5 letter code)"
          },
          {
            "name": "include_synonyms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, include the synonyms of each item in the response.",
              "default": false,
              "title": "Include Synonyms"
            },
            "description": "If true, include the synonyms of each item in the response."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountriesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ingredients": {
      "get": {
        "summary": "Get Ingredients",
        "description": "Get the list of ingredients relevant for green-score computation",
        "operationId": "get_ingredients_v1_ingredients_get",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Language for the request (2 or 5 letter code)",
              "title": "Lang"
            },
            "description": "Language for the request (2 or 5 letter code)"
          },
          {
            "name": "include_synonyms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, include the synonyms of each item in the response.",
              "default": false,
              "title": "Include Synonyms"
            },
            "description": "If true, include the synonyms of each item in the response."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/units": {
      "get": {
        "summary": "Get Units",
        "description": "Get the list of units available in the Open Food Facts units taxonomy\n\nNote: as the list is not too big, we let clients handle suggestions to users",
        "operationId": "get_units_v1_units_get",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Language for the request (2 or 5 letter code)",
              "title": "Lang"
            },
            "description": "Language for the request (2 or 5 letter code)"
          },
          {
            "name": "include_synonyms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, include the synonyms of each item in the response.",
              "default": false,
              "title": "Include Synonyms"
            },
            "description": "If true, include the synonyms of each item in the response."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/suggest-scored-ingredient": {
      "get": {
        "summary": "Suggest Scored Ingredient",
        "description": "Suggest scored ingredient alternatives for an ingredient with no Agribalyse match.\n\nGiven a taxonomy id (typically one returned in ``missing_ingredient_ids``),\nwalk down the ingredients taxonomy and return the descendants that resolve\nto an Agribalyse row, so the user can pick a more specific alternative.",
        "operationId": "suggest_scored_ingredient_v1_suggest_scored_ingredient_get",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Language for the request (2 or 5 letter code)",
              "title": "Lang"
            },
            "description": "Language for the request (2 or 5 letter code)"
          },
          {
            "name": "include_synonyms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, include the synonyms of each item in the response.",
              "default": false,
              "title": "Include Synonyms"
            },
            "description": "If true, include the synonyms of each item in the response."
          },
          {
            "name": "taxonomy_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Taxonomy id of the ingredient to find alternatives for",
              "title": "Taxonomy Id"
            },
            "description": "Taxonomy id of the ingredient to find alternatives for"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestScoredIngredientResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/green-score": {
      "post": {
        "summary": "Green Score",
        "description": "Compute the green-score of a recipe given as a list of ingredients.",
        "operationId": "green_score_v1_green_score_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GreenScoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GreenScoreResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/recompute-quantity": {
      "post": {
        "summary": "Recompute Quantity",
        "description": "Recompute the quantity in grams after the user edited an ingredient's value/unit.\n\nThis is useful to let user change the value of a recipe item in a natural fashion\n(eg. change 1 egg to 3 eggs)\nwhile keeping the equivalent \"g\" conversion for green-score computation.\n\nA best effort is done to also allow changing the unit,\nbut currently, only new units that can be converted to grams are supported",
        "operationId": "recompute_quantity_v1_recompute_quantity_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecomputeQuantityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecomputeQuantityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CountriesResponse": {
        "properties": {
          "countries": {
            "items": {
              "$ref": "#/components/schemas/Country"
            },
            "type": "array",
            "title": "Countries"
          }
        },
        "type": "object",
        "required": [
          "countries"
        ],
        "title": "CountriesResponse",
        "description": "Response model for get_countries endpoint",
        "examples": [
          {
            "countries": [
              {
                "id": "en:france",
                "label": "France"
              },
              {
                "id": "en:spain",
                "label": "Spain"
              }
            ]
          }
        ]
      },
...
      "GreenScoreRequest": {
        "properties": {
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/RecipeIngredientInput"
            },
            "type": "array",
            "title": "Ingredients",
            "description": "The ingredients of the recipe"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country",
            "description": "Country code (ISO 3166-1 alpha-2) to compute the distance modifier for the recipe.If not provided, the distance will always be world"
          },
          "accountedWeights": {
            "$ref": "#/components/schemas/AccountedWeights",
            "description": "Accounted weights for the ponderated sum.\n\n    * scorable takes the ratio of each ingredient compared to the total weight of ingredients that have an EF score\n    * total takes the ratio of each ingredient compared to the total weight of all ingredients, including unmatched ones\n    ",
            "default": "scorable"
          }
        },
        "type": "object",
        "required": [
          "ingredients"
        ],
        "title": "GreenScoreRequest",
        "description": "Request body for the green-score computation endpoint.\n\nIt is a thin wrapper around a list of ingredients",
        "examples": [
          {
            "accountedWeights": "scorable",
            "country": "FR",
            "ingredients": [
              {
                "codifiedIngredient": {
                  "id": "en:apple",
                  "isInTaxonomy": true,
                  "label": "Apple"
                },
                "id": "i1",
                "labels": [],
                "name": "apple",
                "seasonality": false,
                "weight": 150
              },
              {
                "codifiedIngredient": {
                  "id": "en:wheat-flour",
                  "isInTaxonomy": true,
                  "label": "Wheat flour"
                },
                "id": "i2",
                "labels": [],
                "name": "wheat flour",
                "seasonality": false,
                "weight": 200
              }
            ]
          }
        ]
      },
      "GreenScoreResponse": {
        "properties": {
          "globalEfScore": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Globalefscore",
            "description": "The computed global EF score of the recipe, null if no ingredients have a score"
          },
          "labelsBonus": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Labelsbonus",
            "description": "The bonus from ingredient labels, null if no ingredients have a score, 0 if no labels"
          },
          "epiModifier": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Epimodifier",
            "description": "The modifier from ingredient origin agricultural system (EPI), null if no ingredients have a score"
          },
          "distancesModifier": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distancesmodifier",
            "description": "The modifier from ingredient origin distance, null if no ingredients have a score"
          },
          "numericScore": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Numericscore",
            "description": "The computed green-score of the recipe, null if no ingredients have a score"
          },
          "letterGrade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lettergrade",
            "description": "The letter grade corresponding to the numeric score, null if no ingredients have a score"
          },
          "missingIngredientIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missingingredientids",
            "description": "List of ingredient ids that were missing from the Agribalyse computation",
            "default": []
          }
        },
        "type": "object",
        "title": "GreenScoreResponse",
        "description": "Response model for the green-score computation endpoint.",
        "examples": [
          {
            "letterGrade": "A",
            "missingIngredientIds": [],
            "numericScore": 76.38
          }
        ]
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngredientsResponse": {
        "properties": {
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/Ingredient"
            },
            "type": "array",
            "title": "Ingredients"
          }
        },
        "type": "object",
        "required": [
          "ingredients"
        ],
        "title": "IngredientsResponse",
        "description": "Response model for get_ingredients endpoint",
        "examples": [
          {
            "ingredients": [
              {
                "id": "en:apple",
                "label": "Apple"
              },
              {
                "id": "en:wheat-flour",
                "label": "Wheat flour"
              }
            ]
          }
        ]
      },
...
      "LabelsResponse": {
        "properties": {
          "labels": {
            "items": {
              "$ref": "#/components/schemas/Label"
            },
            "type": "array",
            "title": "Labels"
          }
        },
        "type": "object",
        "required": [
          "labels"
        ],
        "title": "LabelsResponse",
        "description": "Response model for get_labels endpoint",
        "examples": [
          {
            "labels": [
              {
                "id": "en:eu-organic",
                "label": "EU Organic"
              },
              {
                "id": "en:fair-trade",
                "label": "Fair Trade"
              }
            ]
          }
        ]
      },
...
      "OriginsResponse": {
        "properties": {
          "origins": {
            "items": {
              "$ref": "#/components/schemas/Origin"
            },
            "type": "array",
            "title": "Origins"
          }
        },
        "type": "object",
        "required": [
          "origins"
        ],
        "title": "OriginsResponse",
        "description": "Response model for get_origins endpoint",
        "examples": [
          {
            "origins": [
              {
                "id": "en:france",
                "label": "France",
                "synonyms": [
                  "french"
                ]
              },
              {
                "id": "en:spain",
                "label": "Spain",
                "synonyms": [
                  "spanish"
                ]
              }
            ]
          }
        ]
      },
...
      "RecipeParseRequest": {
        "properties": {
          "lang": {
            "type": "string",
            "title": "Lang",
            "description": "Language for the request (2 or 5 letter code)"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "lang",
          "text"
        ],
        "title": "RecipeParseRequest",
        "description": "Request model for parse_text endpoint",
        "examples": [
          {
            "lang": "en",
            "text": "200g of apple, 1 cup of wheat flour"
          }
        ]
      },
      "RecipeParseResponse": {
        "properties": {
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/RecipeIngredient"
            },
            "type": "array",
            "title": "Ingredients"
          }
        },
        "type": "object",
        "required": [
          "ingredients"
        ],
        "title": "RecipeParseResponse",
        "description": "Response model for parse_text endpoint",
        "examples": [
          {
            "ingredients": [
              {
                "codified_ingredient": "apple",
                "is_in_taxonomy": true,
                "quantity_g": 150.0,
                "taxonomy_id": "en:apple"
              },
              {
                "codified_ingredient": "wheat flour",
                "is_in_taxonomy": true,
                "quantity_g": 200.0,
                "taxonomy_id": "en:wheat-flour"
              }
            ]
          }
        ]
      },
      "RecomputeQuantityRequest": {
        "properties": {
          "quantityG": {
            "type": "number",
            "minimum": 0.0,
            "title": "Quantityg",
            "description": "Previous quantity in grams"
          },
          "oldValue": {
            "type": "number",
            "minimum": 0.0,
            "title": "Oldvalue",
            "description": "Previous numeric value of the quantity"
          },
          "oldUnit": {
            "type": "string",
            "title": "Oldunit",
            "description": "Previous unit (taxonomy id or 'item')"
          },
          "newValue": {
            "type": "number",
            "minimum": 0.0,
            "title": "Newvalue",
            "description": "New numeric value of the quantity"
          },
          "newUnit": {
            "type": "string",
            "title": "Newunit",
            "description": "New unit (taxonomy id or 'item')"
          }
        },
        "type": "object",
        "required": [
          "quantityG",
          "oldValue",
          "oldUnit",
          "newValue",
          "newUnit"
        ],
        "title": "RecomputeQuantityRequest",
        "description": "Request body for the ``POST /v1/recompute-quantity`` endpoint.\n\n``unit`` is either a unit id from the OFF units taxonomy (e.g. ``xx:kg``)\nor the ``{ITEM_UNIT}`` sentinel for countable ingredients (e.g. \"1 egg\").",
        "examples": [
          {
            "new_unit": "xx:kg",
            "new_value": 2,
            "old_unit": "xx:g",
            "old_value": 2000,
            "quantity_g": 2000
          }
        ]
      },
      "RecomputeQuantityResponse": {
        "properties": {
          "quantityG": {
            "type": "number",
            "title": "Quantityg",
            "description": "New quantity in grams"
          },
          "value": {
            "type": "number",
            "title": "Value",
            "description": "New numeric value of the quantity"
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "description": "New unit (taxonomy id or 'item')"
          }
        },
        "type": "object",
        "required": [
          "quantityG",
          "value",
          "unit"
        ],
        "title": "RecomputeQuantityResponse",
        "description": "Response model for the ``POST /v1/recompute-quantity`` endpoint.",
        "examples": [
          {
            "quantity_g": 2000,
            "unit": "xx:kg",
            "value": 2
          }
        ]
      },
...
      "SuggestScoredIngredientResponse": {
        "properties": {
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/ScoredIngredient"
            },
            "type": "array",
            "title": "Ingredients"
          }
        },
        "type": "object",
        "required": [
          "ingredients"
        ],
        "title": "SuggestScoredIngredientResponse",
        "description": "Response model for the suggest-scored-ingredient endpoint."
      },
...
      "UnitsResponse": {
        "properties": {
          "units": {
            "items": {
              "$ref": "#/components/schemas/Unit"
            },
            "type": "array",
            "title": "Units"
          }
        },
        "type": "object",
        "required": [
          "units"
        ],
        "title": "UnitsResponse",
        "description": "Response model for get_units endpoint",
        "examples": [
          {
            "units": [
              {
                "id": "en:gram",
                "label": "gram",
                "standard_unit": "g"
              },
              {
                "id": "en:cup",
                "label": "cup",
                "standard_unit": "ml"
              }
            ]
          }
        ]
      },
...
    }
  }
}